rename trait Adjustable -> Controller
