+++ src/frame/capturer/wayland.rs
@@ -1,7 +1,7 @@
 use crate::config::WaylandProtocol;
 use crate::frame::object::Object;
 use crate::frame::vulkan::Vulkan;
-use crate::frame::capturer::Adjustable;
+use crate::predictor::Controller;
 use std::os::fd::BorrowedFd;
 use std::thread;
 use std::time::Duration;
@@ -36,7 +36,7 @@ pub struct Capturer {
     output: Option<WlOutput>,
     output_global_id: Option<u32>,
     pending_frame: Option<Object>,
-    controller: Option<Box<dyn Adjustable>>,
+    controller: Option<Box<dyn Controller>>,
     // linux-dmabuf-v1
     dmabuf: Option<ZwpLinuxDmabufV1>,
     wl_buffer: Option<WlBuffer>,
@@ -84,7 +84,7 @@ impl Capturer {
 }
 
 impl super::Capturer for Capturer {
-    fn run(&mut self, output_name: &str, controller: Box<dyn Adjustable>) {
+    fn run(&mut self, output_name: &str, controller: Box<dyn Controller>) {
         let connection =
             Connection::connect_to_env().expect("Unable to connect to Wayland display");
         let display = connection.display();
