+++ src/predictor/controller/smart.rs
@@ -2,9 +2,9 @@ use crate::predictor::data::{Data, Entry};
 use std::sync::mpsc::{Receiver, Sender};
 use std::time::Duration;
 
-use crate::frame::capturer::Adjustable;
-
-use super::{INITIAL_TIMEOUT_SECS, NEXT_ALS_COOLDOWN_RESET, PENDING_COOLDOWN_RESET};
+use super::{
+    Controller as _, INITIAL_TIMEOUT_SECS, NEXT_ALS_COOLDOWN_RESET, PENDING_COOLDOWN_RESET,
+};
 
 pub struct Controller {
     prediction_tx: Sender<u64>,
@@ -21,7 +21,7 @@ pub struct Controller {
     output_name: String,
 }
 
-impl Adjustable for Controller {
+impl super::Controller for Controller {
     fn adjust(&mut self, luma: u8) {
         if self.last_als.is_none() {
             // ALS controller is expected to send the initial value on this channel asap
