+++ src/predictor/mod.rs
@@ -1,2 +1,3 @@
 pub mod controller;
 pub mod data;
+pub use controller::Controller;
